Hide some spell mechnaics details
